leaderboard-attendance-svc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
psycopg[binary]==3.2.10

pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.5.2

httpx==0.27.2
PyJWT==2.9.0

prometheus-fastapi-instrumentator==7.1.0
nats-py==2.6.0
APScheduler==3.10.4

python-dotenv==1.0.1
pandas==2.2.3
python-jose==3.3.0
aiosqlite==0.20.0
pytest==8.3.3
pytest-asyncio==0.24.0
requests==2.32.3